Added sideband detection to FreeDV's Hamlib support.
Per agreed convention, FreeDV uses LSB for bands under 10MHz and USB for bands above it.
This differs from the usual "data mode" convention of USB regardless of band, This commit
adds the ability for FreeDV to detect the sideband currently being used by the radio and
emit a warning dialog if it differs from the expected convention.

Limitations:

1. Hamlib support must be enabled in FreeDV's settings.
2. Detection is only done when the user pushes Start. If the sideband changes during the
   session, FreeDV will not detect the change.
3. FreeDV/Codec2 will not attempt to mitigate incorrect sideband settings during the
   decode process.