Convert samples from voice keyer file to sound card 2's sample rate.
When sound card 2's sample rate is less than sound card 1's, FreeDV terminates
transmit too early when using the voice keyer even if the .wav file is in the
mode's expected sample rate (e.g. 8 or 16KHz). This change ensures that the
samples from the voice keyer match the sample rate that has been configured by
the user prior to being passed to Codec2.

This resolves #48 and potentially also #144.