Fix bug preventing proper Options window sizing on Windows. (#478)
Reported by several people from the digitalvoice mailing list, it appears that the OK/Cancel/Apply buttons are now hidden on 1.8.12 in Windows unless the window is resized. Additionally, I've noticed some oddities with the frequency list control sizing when the Options window's resized on Linux (tested on Ubuntu 22.04 LTS). This PR does the following:

1. Disables resizing of the Options window. Since the other controls on that dialog don't resize, there's not much point in allowing users to do so.
2. Forces sizing of the various frequency list related controls to prevent them from being resized when data is loaded from the user's configuration.